Move things to more generic location
