+++ sfm_pc/sql/areas.sql
