+++ sfm_pc/urls.py
@@ -8,7 +8,7 @@ from django.views.decorators.cache import cache_page
 from django.core.urlresolvers import reverse_lazy
 
 from sfm_pc.views import (Dashboard, osm_autocomplete, division_autocomplete,
-                          Countries, command_chain,
+                          Countries, command_chain, DownloadData,
                           download_zip, Help, About, country_background)
 
 urlpatterns = i18n_patterns(
@@ -22,6 +22,7 @@ urlpatterns = i18n_patterns(
     url(r'^help/', Help.as_view(), name="help"),
     url(r'^about/', About.as_view(), name="about"),
     url(r'^search/', include('search.urls')),
+    url(r'^download/', DownloadData.as_view(), name="download"),
 
     url(r'^osm-autocomplete/$', osm_autocomplete, name="osm-autocomplete"),
     url(r'^division-autocomplete/$', division_autocomplete, name="division-autocomplete"),
