+++ templates/download.html
@@ -2,7 +2,7 @@
 {% load i18n %}
 {% block content %}
     <h2>{% trans "Download information about organizations" %}</h2>
-    <form action='{% url 'organization-download' %}' method='POST'>
+    <form action='{% url 'download' %}' method='POST'>
         {% csrf_token %}
         <div class="row">
             <div class="col-sm-12">
